# Add bsf(ulong), bsr(ulong), and popcnt(ulong).