This PR adds `ulong` versions of `bsf()`, `bsr()`, and `popcnt()`. (It also fixes a small error in the `_popcnt()` docs.)

I use these `core.bitop` functions fairly often, and am always annoyed that the maximum bit-ness of the input depends on what architecture I'm compiling for; it makes using these ops from generic code painful.

There is no good reason for this, as even on 32-bit systems the 64-bit op can be trivially implemented in terms of a couple of calls to the 32-bit op.

**EDIT:** I have created a [companion PR for Phobos](https://github.com/D-Programming-Language/phobos/pull/3888). While working on it, I discovered two different private implementations of `bsr()` for `ulong`, one of which is bugged.