@tsbockman 2015-12-28
I was able to massively speed up `popcnt()` in most circumstances by:

1. Generalizing the old software algorithm to directly cover 64-bit as well. (Roughly x+3 ops for the 64-bit version; much better than the naive (2*x)+1 ops.)
2. Only using 64-bit ops if `size_t.sizeof == ulong.sizeof`.
3. On DMD, use `_popcnt()` if it is available. There is a small performance penalty associated with checking `core.cpuid.hasPopcnt`, but a **massive** speed-up if it returns `true`, allowing the use of the SSE4 `POPCNT` instruction.