@ibuclaw 2015-12-28
Considering this comment: https://github.com/D-Programming-Language/druntime/blob/4ed3b721cc9c6b99f9400817853e430487e3d45b/src/core/checkedint.d#L18-L20

I don't know why @WalterBright introduced ``_popcnt``.  Would have been better off recognising ``popcnt`` as an intrinsic, and ignoring the body.