@tsbockman 2015-12-28
> I don't know why @WalterBright introduced `_popcnt`.

I suspect it's related to the fact that you can't count on every CPU having SSE4 support. The other `core.bitop` intrinsics should work on any x86 CPU supported by D, right?