@tsbockman 2015-12-28
I have tested the new `popcnt()` algorithm in commit `b460981b06b51d6756e742d405bafa139a5d6890` exhaustively against the old one with the following:
```D
void verify32(string funStr)() {
    mixin("alias fun = " ~ funStr ~ ";");
    uint badCnt = 0;

    uint i = 0;
    while(true) {
        if((i & 0xFFF_FFFF) == 0) {
            write(i >>> 28);
            write(';');
            stdout.flush();
        }

        if(popcnt(i) != fun(i))
            ++badCnt;
        if(i == uint.max)
            break;
        ++i;
    }

    writefln("\r\n" ~ funStr ~ " fails: %s", badCnt);
}
void verify64(string funStr)() {
    mixin("alias fun = " ~ funStr ~ ";");
    ulong badCnt = 0uL;

    for(ulong i = 0; i <= uint.max; ++i)
    {
        if((i & 0xFFF_FFFF) == 0) {
            write(i >>> 28);
            write(';');
            stdout.flush();
        }

        const pi = popcnt(cast(uint)i);
        if(pi != fun(i))
            ++badCnt;

        const ib = i << 12;
        if(pi != fun(ib))
            ++badCnt;

        const ic = ~(i << 17);
        if((64 - pi) != fun(ic))
            ++badCnt;

        const id = Split64(ib ^ ic);
        if((popcnt(id.lo) + popcnt(id.hi)) != fun(id.u64))
            ++badCnt;
    }

    writefln("\r\n" ~ funStr ~ " fails: %s", badCnt);
}
```
Passed so far (all on an Intel Haswell running linux): DMD (64- and 32-bit)