@tsbockman 2015-12-28
Micro-benchmark results:
```
// DMD 64-bit:
SSE4 `_popcnt(uint)` :   82 ms
Old   `popcnt(uint)` : 5902 ms
New   `popcnt(uint)` :  205 ms

SSE4 `_popcnt(ulong)`:  147 ms
New   `popcnt(ulong)`:  179 ms

// DMD 32-bit:
SSE4 `_popcnt(uint)` :   83 ms
Old   `popcnt(uint)` : 7677 ms
New   `popcnt(uint)` :  126 ms

New   `popcnt(ulong)`: 2403 ms

// GDC 64-bit:
Old   `popcnt(uint)` : 2369 ms
New   `popcnt(uint)` : 1711 ms
New   `popcnt(ulong)`: 2282 ms

// GDC 32-bit:
Old   `popcnt(uint)` : 4874 ms
New   `popcnt(uint)` : 1632 ms
New   `popcnt(ulong)`: 5056 ms

// LDC 64-bit:
Old   `popcnt(uint)` : 1844 ms
New   `popcnt(uint)` : 1904 ms
New   `popcnt(ulong)`: 2404 ms
```
The error bars on these numbers are huge for various reasons, but hopefully this is sufficient to demonstrate that the new algorithm is an improvement.