@ibuclaw 2015-12-28
> // GDC 64-bit:
> Old   `popcnt(uint)` : 2369 ms
> New   `popcnt(uint)` : 1711 ms
> New   `popcnt(ulong)`: 2282 ms
>
> // GDC 32-bit:
> Old   `popcnt(uint)` : 4874 ms
> New   `popcnt(uint)` : 1632 ms
> New   `popcnt(ulong)`: 5056 ms

Challenge accepted. :-)