@tsbockman 2015-12-30
> Why not use multiply method?

I benchmarked your method. On my machine (Intel Haswell), it's precisely identical in speed for `ulong`, and 1.3-4.9x faster for `uint`. (I have also verified correctness using the same test as before.)

So, this PR is now strictly faster on my machine than the old implementation, even without using the SSE4 `POPCNT`; the small overhead of checking `core.cpuid.hasPopcnt` is overcome by the improved `soft_popcnt!uint()`.