@tsbockman 2015-12-30
The auto-tester failure is unrelated to my PR:
```
Exception - file 'lexer.h': lexer.h: No such file or directory
Exception - file 'parse.h': parse.h: No such file or directory
Error - file 'doc.d' contains trailing whitespace at line 2526
Exception - file 'gluestub.d': gluestub.d: No such file or directory
```