@tsbockman 2016-01-11
@ibuclaw Does this need more work, or would you say it's ready if I can get a second reviewer to approve it?