@andralex 2016-01-12
This adds yet another punctuation style to the several ones we have... please stick with space after `if`, follow the existing style for `else static if` etc.