@tsbockman 2016-01-12
> This adds yet another punctuation style to the several ones we have... please stick with space after `if`, follow the existing style for else `static if` etc.

@andralex If you want to enforce arbitrary requirements for semantically irrelevant stuff like whitespace usage, please add them to the [official style guide](http://dlang.org/dstyle.html) or the [Phobos contribution guidelines](http://wiki.dlang.org/Phobos_Quality).

I have updated the `if` and `static if` stuff like you asked, but I'm not even going to try to guess what all "**etc.**" might encompass since, as you yourself admitted, there is no consistent "existing style" to follow.