@JakobOvrum 2016-01-12
@tsbockman, filed style update: D-Programming-Language/dlang.org/pull/1191