@tsbockman 2016-01-12
> filed style update: [D-Programming-Language/dlang.org#1191](https://github.com/D-Programming-Language/dlang.org/pull/1191)

My changes should meet the whitespace requirements now. (Unlike the rest of the file...)