@andralex 2016-01-12
OK. Shall we return `uint` when there's no possibility of a negative?