@tsbockman 2016-01-12
> Shall we return `uint` when there's no possibility of a negative?

That would be a breaking change:
* Someone might be relying on the current return type by capturing it with `auto`.
* Stuff like [Issue 259](https://issues.dlang.org/show_bug.cgi?id=259) makes changing API types between signed and unsigned *very* dangerous.

(I've heard `null` called "*the billion dollar mistake*", but is it really worse than this?
```D
int  a = -1;
uint b =  1;
assert(a > b);
```
At least trying to dereference `null` usually generates a segfault or something. And unlike `null`, the fundamental fix is so simple, regardless of whether you implement it in hardware or software...)