@tsbockman 2016-01-12
More generally, because of the way that integer type promotion works currently, *outputs* should always be signed, if their range is guaranteed to fit in an appropriate signed type. Mixed signed/unsigned operations are always safe - well, as safe as same-signedness operations, anyway - provided the signed value is provably positive.

Unfortunately, I don't think there is any safe default choice for *inputs* with the current fail-silently-as-fast-as-possible rules.