@andralex 2016-01-13
Auto-merge toggled on