@tsbockman 2016-01-14
@andralex Since this is merged now, please consider taking a look at the companion PR for Phobos: https://github.com/D-Programming-Language/phobos/pull/3888