@tsbockman 2016-01-21
I broke `popcnt()` in CTFE with this PR. Fix: https://github.com/D-Programming-Language/druntime/pull/1475