@denis-sh 2016-01-28
@tsbockman, `size_t` isn't some "magic" type in D, it's just an `alias` for a corresponding integer type so `int bsf(size_t v)` is actually `int bsf(ulong v)` if pointers are 64 bits. As a result `bsf(cast(size_t) v)` is equal to `bsf(v)` if `size_t.sizeof == ulong.sizeof` so your function implementation recursively calls itself.
Opened fixup pull #1486.

##### P.S.
Looks like we don't have enough human resources to carefully review pull requests.