@tsbockman 2016-01-28
> your function implementation recursively calls itself.

This claim confused me greatly, until I saw [the explanation you posted elsewhere](https://github.com/D-Programming-Language/druntime/pull/1486#issue-129376824):
> It didn't cause tester failure because dmd uses intrinsic even if there is a function implementation.